Use only the hard exit LeaderElectionSpec, see #2113
